Finishing up support for special attribute fields
