+++ lib/graphql/activerecord.rb
@@ -14,5 +14,6 @@ require 'graphql/models/identification/model_types'
 require 'graphql/models/identification/node_identification'
 require 'graphql/models/proxy_block'
 require 'graphql/models/definition_helpers'
+require 'graphql/models/definition_helpers/associations'
 require 'graphql/models/model_type_config'
 require 'graphql/models/model_type'
